Minimize testutil files
